requirements.txt
tensorflow
keras
opencv-python
pillow
numpy
pywin32